column

+++ be/src/column/array_column.cpp
@@ -216,7 +216,7 @@ void ArrayColumn::remove_first_n_values(size_t count) {
         count = _offsets->size() - 1;
     }
 
-    size_t offset = _offsets->get_data()[count];
+    uint32_t offset = _offsets->get_data()[count];
     _elements->remove_first_n_values(offset);
     _offsets->remove_first_n_values(count);
